+++ src/frontend/admin/src/services/factories/orders/index.ts
@@ -82,6 +82,8 @@ const build = (state?: OrderStatesEnum): Order => {
     ],
     credit_card: CreditCardFactory(),
     has_waived_withdrawal_right: faker.datatype.boolean(),
+    withdrawn_requested_at: null,
+    withdrawn_confirmation_at: null,
     voucher: null,
   };
   if (
@@ -136,6 +138,7 @@ const buildOrderListItem = (): OrderListItem => {
     total: faker.number.float({ min: 1, max: 9999 }),
     total_currency: "EUR",
     voucher: null,
+    withdrawn_confirmation_at: null,
   };
 };